requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-import-export==2.5.0
django-prettyjson==0.4.1
et-xmlfile==1.1.0
idna==3.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
pretty-json==1.2.0
psycopg2==2.8.6
Pygments==2.10.0
pygments-json==0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
standardjson==0.3.1
tablib==3.0.0
urllib3==1.26.6
xlrd==2.0.1
xlwt==1.3.0
